Use list of F-Chat protocall commands
This adds a full, descriptive list of F-Chat commands to the 'fchat.py' module. The aim of this is twofold. Firstly it improves readablity, for example `message(CREATE_PRIVATE_CHANNEL, {'channel':channelname})` is more readable and obvious than `message("CRR", {'channel':channelname})`. Secondly it means that if a command is ever changed we only have to update it in a single place, and applications using this library don't have to change anything in their code.